chore: move fixtures and helpers to root (#11812)
