+++ superset-frontend/spec/fixtures/mockDashboardInfo.js
