+++ superset-frontend/spec/javascripts/dashboard/reducers/dashboardFilters_spec.js
@@ -30,13 +30,13 @@ import * as activeDashboardFilters from 'src/dashboard/util/activeDashboardFilte
 import {
   emptyFilters,
   dashboardFilters,
-} from '../fixtures/mockDashboardFilters';
+} from 'spec/fixtures/mockDashboardFilters';
 import {
   sliceEntitiesForDashboard,
   filterId,
   column,
-} from '../fixtures/mockSliceEntities';
-import { filterComponent } from '../fixtures/mockDashboardLayout';
+} from 'spec/fixtures/mockSliceEntities';
+import { filterComponent } from 'spec/fixtures/mockDashboardLayout';
 
 describe('dashboardFilters reducer', () => {
   const { form_data } = sliceEntitiesForDashboard.slices[filterId];
