+++ superset-frontend/spec/javascripts/explore/components/FilterBoxItemControl_spec.jsx
@@ -24,7 +24,7 @@ import { shallow } from 'enzyme';
 import Popover from 'src/common/components/Popover';
 import FilterBoxItemControl from 'src/explore/components/controls/FilterBoxItemControl';
 import FormRow from 'src/components/FormRow';
-import datasources from '../../../fixtures/mockDatasource';
+import datasources from 'spec/fixtures/mockDatasource';
 
 const defaultProps = {
   label: 'some label',
